requirements.txt
streamlit
python-dotenv
google-genai
huggingface_hub
Pillow
requests